Refactor: move `getSearchDetailsForUserById` into a search helper file
