+++ src/infrastructure/search/api.js
@@ -1,7 +1,6 @@
 const jwtStrategy = require("login.dfe.jwt-strategies");
 const config = require("./../config");
 const { fetchApi } = require("login.dfe.async-retry");
-const { searchUserByIdRaw } = require("login.dfe.api-client/users");
 const { mapSearchUserToSupportModel } = require("../../app/users/utils");
 
 const callApi = async (endpoint, method, body, correlationId) => {
@@ -77,15 +76,6 @@ const searchForUsers = async (
   }
 };
 
-const getSearchDetailsForUserById = async (id) => {
-  try {
-    const user = await searchUserByIdRaw({ userId: id });
-    return user ? mapSearchUserToSupportModel(user) : undefined;
-  } catch (e) {
-    throw new Error(`Error getting user ${id} from search - ${e.message}`);
-  }
-};
-
 const getById = async (userId, correlationId) => {
   const token = await jwtStrategy(config.search.service).getBearerToken();
   try {
@@ -127,7 +117,6 @@ const createIndex = async (id, correlationId) => {
 
 module.exports = {
   searchForUsers,
-  getSearchDetailsForUserById,
   getById,
   updateUserInSearch,
   updateIndex,
