+++ test/app/accessRequests/reviewOrganisationRequest.post.test.js
@@ -10,6 +10,9 @@ jest.mock("./../../../src/infrastructure/organisations");
 jest.mock("./../../../src/infrastructure/search");
 jest.mock("login.dfe.jobs-client");
 jest.mock("login.dfe.api-client/organisations");
+jest.mock(
+  "../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById",
+);
 
 const { getRequestMock, getResponseMock } = require("./../../utils");
 const {
@@ -21,9 +24,7 @@ const {
   putUserInOrganisation,
   updateRequestById,
 } = require("./../../../src/infrastructure/organisations");
-const {
-  getSearchDetailsForUserById,
-} = require("./../../../src/infrastructure/search");
+
 const {
   getAndMapOrgRequest,
 } = require("./../../../src/app/accessRequests/utils");
@@ -32,6 +33,9 @@ const { NotificationClient } = require("login.dfe.jobs-client");
 const {
   getOrganisationLegacyRaw,
 } = require("login.dfe.api-client/organisations");
+const {
+  getSearchDetailsForUserById,
+} = require("../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById");
 
 const sendAccessRequest = jest.fn();
 NotificationClient.mockImplementation(() => ({
