+++ test/app/accessRequests/selectPermissionLevel.post.test.js
@@ -10,6 +10,9 @@ jest.mock("./../../../src/infrastructure/organisations");
 jest.mock("./../../../src/infrastructure/search");
 jest.mock("login.dfe.jobs-client");
 jest.mock("login.dfe.api-client/organisations");
+jest.mock(
+  "../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById",
+);
 
 const { getRequestMock, getResponseMock } = require("./../../utils");
 const {
@@ -21,10 +24,7 @@ const {
   putUserInOrganisation,
   updateRequestById,
 } = require("./../../../src/infrastructure/organisations");
-const {
-  getSearchDetailsForUserById,
-  updateIndex,
-} = require("./../../../src/infrastructure/search");
+const { updateIndex } = require("./../../../src/infrastructure/search");
 const {
   getAndMapOrgRequest,
 } = require("./../../../src/app/accessRequests/utils");
@@ -39,6 +39,9 @@ NotificationClient.mockImplementation(() => ({
 const {
   getOrganisationLegacyRaw,
 } = require("login.dfe.api-client/organisations");
+const {
+  getSearchDetailsForUserById,
+} = require("../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById");
 
 Date.now = jest.fn(() => "2019-01-02");
