+++ test/app/users/postEditPermissions.test.js
@@ -7,6 +7,9 @@ jest.mock("./../../../src/infrastructure/logger", () =>
 jest.mock("./../../../src/infrastructure/organisations");
 jest.mock("./../../../src/infrastructure/search");
 jest.mock("login.dfe.api-client/invitations");
+jest.mock(
+  "../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById",
+);
 
 const { getRequestMock, getResponseMock } = require("./../../utils");
 const postEditPermissions = require("./../../../src/app/users/postEditPermissions");
@@ -14,9 +17,6 @@ const {
   setUserAccessToOrganisation,
   getUserOrganisations,
 } = require("./../../../src/infrastructure/organisations");
-const {
-  getSearchDetailsForUserById,
-} = require("./../../../src/infrastructure/search");
 
 jest.mock("login.dfe.jobs-client");
 jest.mock("login.dfe.api-client/services");
@@ -24,6 +24,9 @@ const { NotificationClient } = require("login.dfe.jobs-client");
 const {
   addOrganisationToInvitation,
 } = require("login.dfe.api-client/invitations");
+const {
+  getSearchDetailsForUserById,
+} = require("../../../src/app/users/searchApiHelpers/getSearchDetailsForUserById");
 
 const res = getResponseMock();
